Time: 71 ms (50.00%), Space: 101.9 MB (50.00%) - LeetHub

+++ 2195-append-k-integers-with-minimal-sum/2195-append-k-integers-with-minimal-sum.java
@@ -5,7 +5,7 @@ class Solution {
         HashSet<Integer> set=new HashSet<>();
         for(int num:nums){
             if(num<=k && !set.contains(num)){
-                sum=sum-num;
+                sum=sum-(long)num;
                 count++;
             }
             set.add(num);
